+++ pricing/src/epilot/sdkconfiguration.py
@@ -20,8 +20,8 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '1.0.0'
-    sdk_version: str = '1.40.3'
-    gen_version: str = '2.108.3'
+    sdk_version: str = '1.40.4'
+    gen_version: str = '2.115.2'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: